fix: bug fix for PR and modularized component
