+++ src/pages/salesman/MeetingsManagement/MeetingDetail/MeetingDetail.tsx
@@ -1,69 +1,36 @@
-import { EPageRoutes } from '@data/enums/EPageRoutes';
-import { getSentimentConfig } from '@hooks/useSentiment';
-import { ArrowBack } from '@mui/icons-material';
-import EventNoteOutlinedIcon from '@mui/icons-material/EventNoteOutlined';
 import {
   Box,
   CircularProgress,
-  type Color,
-  IconButton,
-  Link as MuiLink,
   Stack,
-  Tab,
-  Tabs,
-  Typography,
   useMediaQuery,
   useTheme,
 } from '@mui/material';
 import { PageNotFound } from '@pages/PageNotFound/PageNotFound';
-import { MeetingListItemApiSchema } from '@services/models/MeetingSchema';
 import { useGetMeetingById } from '@services/queries/useMeetings';
-import {
-  Link,
-  useNavigate,
-  useParams,
-  useSearch,
-} from '@tanstack/react-router';
+import { useNavigate, useParams, useSearch } from '@tanstack/react-router';
 import { PageContainter } from '@UI/PageContainer/PageContainer';
-import { StatCard } from '@UI/StatCard/StatCard';
 import { type JSX } from 'react';
 import React from 'react';
-import { z } from 'zod';
 
-// Type adicional para o overall_sentiment, não sei se o adiciono diretamente no Schema de volta.
-type TMeetingDetail = z.infer<typeof MeetingListItemApiSchema> & {
-  client: { overall_sentiment?: number };
-};
-
-// Placeholders
-const MeetingContext = (): JSX.Element => (
-  <Box p={3}>
-    <Typography>Conteúdo: Contexto da Reunião</Typography>
-  </Box>
-);
-const MeetingInsights = (): JSX.Element => (
-  <Box p={3}>
-    <Typography>Conteúdo: Insights na Reunião</Typography>
-  </Box>
-);
-const MeetingActionPlan = (): JSX.Element => (
-  <Box p={3}>
-    <Typography>Conteúdo: Plano de Ação</Typography>
-  </Box>
-);
+import { MeetingDetailHeaderStats } from './components/MeetingDetailHeaderStats';
+import { MeetingDetailTabsContent } from './components/MeetingDetailTabsContent';
+import type { TMeetingDetail, TMeetingTab } from './MeetingDetail.interface';
 
 export const MeetingDetail = (): JSX.Element => {
   const { palette, breakpoints } = useTheme();
   const isDesktop = useMediaQuery(breakpoints.up('lg'));
 
-  const themePalette = palette as typeof palette &
-    Record<string, Partial<Color>>;
+  const themePalette = palette as typeof palette;
 
   const { meetingId } = useParams({ strict: false }) as { meetingId: string };
-  const search = useSearch({ strict: false }) as Record<string, string>;
+  const search = useSearch({
+    from: '/protected/vendedores/reuniões/$meetingId',
+  });
 
-  const currentTab = search.tab || 'context';
-  const navigate = useNavigate();
+  const currentTab = (search.tab || 'context') as TMeetingTab;
+  const navigate = useNavigate({
+    from: '/vendedores/reuniões/$meetingId',
+  });
 
   const {
     data: rawMeeting,
@@ -78,7 +45,10 @@ export const MeetingDetail = (): JSX.Element => {
     newValue: string
   ): void => {
     navigate({
-      search: (old: Record<string, string>) => ({ ...old, tab: newValue }),
+      search: ((old: Record<string, string>) => ({
+        ...old,
+        tab: newValue as TMeetingTab,
+      })) as never,
       replace: true,
     });
   };
@@ -102,168 +72,24 @@ export const MeetingDetail = (): JSX.Element => {
     return <PageNotFound />;
   }
 
-  const formattedDate = meeting.scheduled_for
-    ? new Date(meeting.scheduled_for).toLocaleDateString('pt-BR')
-    : '';
-  const durationMinutes = meeting.duration_seconds
-    ? Math.floor(meeting.duration_seconds / 60)
-    : 0;
-
   const responsiveValueFontSize = isDesktop
     ? '1.25rem !important'
     : '1rem !important';
 
-  const sentimentConfig = getSentimentConfig(meeting.client.overall_sentiment);
-  const sentimentPalette = themePalette[sentimentConfig.theme];
-
   return (
     <PageContainter>
       <Stack spacing={3} sx={{ width: '100%', alignSelf: 'flex-start' }}>
-        <MuiLink
-          component={Link}
-          to={EPageRoutes.SALESMAN_MEETINGS}
-          underline="hover"
-          sx={{
-            display: 'inline-flex',
-            alignItems: 'center',
-            gap: 0.75,
-            color: palette.primary[600],
-            fontWeight: 500,
-            width: 'fit-content',
-            fontSize: '0.875rem',
-          }}
-        >
-          <ArrowBack sx={{ fontSize: '0.875rem' }} />
-          Voltar para reuniões
-        </MuiLink>
+        <MeetingDetailHeaderStats
+          meeting={meeting}
+          palette={palette}
+          themePalette={themePalette}
+          responsiveValueFontSize={responsiveValueFontSize}
+        />
 
-        <Stack direction="row" alignItems="center" spacing={2} flexWrap="wrap">
-          <IconButton
-            disableRipple
-            sx={{
-              width: 48,
-              height: 48,
-              p: 0,
-              borderRadius: '0.75rem',
-              bgcolor: palette.meetings[200],
-              color: palette.meetings[500],
-              '&:hover': {
-                bgcolor: palette.meetings[100],
-              },
-            }}
-          >
-            <EventNoteOutlinedIcon sx={{ fontSize: '1.5rem' }} />
-          </IconButton>
-
-          <Stack direction="column" spacing={0.5} sx={{ minWidth: 0 }}>
-            <Typography
-              variant="h1"
-              component="h1"
-              sx={{ wordBreak: 'break-word', fontSize: '1.5rem' }}
-            >
-              {meeting.title}
-            </Typography>
-            <Typography variant="body2" color="text.secondary">
-              {formattedDate}{' '}
-              {durationMinutes > 0 && `- ${durationMinutes} min`}
-            </Typography>
-          </Stack>
-        </Stack>
-
-        <Stack
-          direction={{ xs: 'column', sm: 'row' }}
-          spacing={2}
-          useFlexGap
-          sx={{ width: '100%', alignItems: 'stretch' }}
-        >
-          <StatCard
-            iconName="real_estate_agent"
-            theme="primary"
-            value={meeting.client.name}
-            label="Nome do cliente"
-            sx={{
-              flex: '1 1 0',
-              minWidth: { xs: '100%', sm: 160 },
-              '& .MuiTypography-root:nth-of-type(1)': {
-                fontSize: responsiveValueFontSize,
-              },
-            }}
-          />
-          <StatCard
-            iconName="salesman"
-            theme="salesmen"
-            value={meeting.seller.name}
-            label="Vendedor responsável"
-            sx={{
-              flex: '1 1 0',
-              minWidth: { xs: '100%', sm: 160 },
-              '& .MuiTypography-root:nth-of-type(1)': {
-                fontSize: responsiveValueFontSize,
-              },
-            }}
-          />
-          <StatCard
-            iconName="company"
-            theme="companies"
-            value={meeting.client.client_company_name}
-            label="Empresa responsável"
-            sx={{
-              flex: '1 1 0',
-              minWidth: { xs: '100%', sm: 160 },
-              '& .MuiTypography-root:nth-of-type(1)': {
-                fontSize: responsiveValueFontSize,
-              },
-            }}
-          />
-          <StatCard
-            iconName={sentimentConfig.iconName}
-            theme={sentimentConfig.theme}
-            value={`${meeting.client.overall_sentiment ?? 0}%`}
-            label="Sentimento da reunião"
-            sx={{
-              flex: '1 1 0',
-              minWidth: { xs: '100%', sm: 160 },
-              '& .MuiTypography-root:nth-of-type(1)': {
-                fontSize: responsiveValueFontSize,
-              },
-              '& .MuiStack-root > :first-child': {
-                backgroundColor: sentimentPalette?.[100],
-                color: sentimentPalette?.[300],
-              },
-              '& .MuiStack-root > :first-child svg': {
-                color: sentimentPalette?.[300],
-              },
-            }}
-          />
-        </Stack>
-
-        <Box
-          sx={{
-            bgcolor: 'background.paper',
-            borderRadius: '1rem',
-            border: '1px solid',
-            borderColor: 'divider',
-            overflow: 'hidden',
-          }}
-        >
-          <Box sx={{ borderBottom: 1, borderColor: 'divider' }}>
-            <Tabs
-              value={currentTab}
-              onChange={handleTabChange}
-              variant="fullWidth"
-            >
-              <Tab label="CONTEXTO DA REUNIÃO" value="context" />
-              <Tab label="INSIGHTS NA REUNIÃO" value="insights" />
-              <Tab label="PLANO DE AÇÃO" value="plan" />
-            </Tabs>
-          </Box>
-
-          <Box minHeight="300px">
-            {currentTab === 'context' && <MeetingContext />}
-            {currentTab === 'insights' && <MeetingInsights />}
-            {currentTab === 'plan' && <MeetingActionPlan />}
-          </Box>
-        </Box>
+        <MeetingDetailTabsContent
+          currentTab={currentTab}
+          onTabChange={handleTabChange}
+        />
       </Stack>
     </PageContainter>
   );
